+++ benchmark/ec/ec_double_benchmark_gpu.cc
@@ -23,14 +23,14 @@ namespace {
 // TODO(chokobole): Use openmp.
 void TestDoubleOnCPU(const std::vector<math::bn254::G1AffinePoint>& bases,
                      std::vector<math::bn254::G1JacobianPoint>& results,
-                     uint64_t nums) {
-  for (uint64_t i = 0; i < nums; ++i) {
+                     size_t nums) {
+  for (size_t i = 0; i < nums; ++i) {
     results[i] = bases[i].Double();
   }
 }
 
 gpuError_t LaunchDouble(const math::bn254::G1AffinePointGpu* x,
-                        math::bn254::G1JacobianPointGpu* y, uint64_t count) {
+                        math::bn254::G1JacobianPointGpu* y, size_t count) {
   math::kernels::Double<<<(count - 1) / 32 + 1, 32>>>(x, y, count);
   gpuError_t error = LOG_IF_GPU_LAST_ERROR("Failed Double()");
   return error == gpuSuccess ? LOG_IF_GPU_ERROR(gpuDeviceSynchronize(),
@@ -41,8 +41,8 @@ gpuError_t LaunchDouble(const math::bn254::G1AffinePointGpu* x,
 void TestDoubleOnGPU(math::bn254::G1AffinePointGpu* bases_cuda,
                      math::bn254::G1JacobianPointGpu* results_cuda,
                      const std::vector<math::bn254::G1AffinePoint>& bases,
-                     uint64_t nums) {
-  for (uint64_t i = 0; i < nums; ++i) {
+                     size_t nums) {
+  for (size_t i = 0; i < nums; ++i) {
     bases_cuda[i] = ConvertPoint<math::bn254::G1AffinePointGpu>(bases[i]);
   }
 
@@ -60,15 +60,15 @@ int RealMain(int argc, char** argv) {
   math::bn254::G1Curve::Init();
   math::bn254::G1CurveGpu::Init();
 
-  const std::vector<uint64_t>& point_nums = config.point_nums();
+  const std::vector<size_t>& point_nums = config.point_nums();
   SimpleReporter reporter("EC double benchmark");
 
   reporter.SetXLabel("# of points");
   reporter.SetColumnLabels(base::Map(
-      point_nums, [](uint64_t num) { return base::NumberToString(num); }));
+      point_nums, [](size_t num) { return base::NumberToString(num); }));
 
   std::cout << "Generating random points..." << std::endl;
-  uint64_t max_point_num = point_nums.back();
+  size_t max_point_num = point_nums.back();
   std::vector<bn254::G1AffinePoint> bases =
       CreatePseudoRandomPoints<bn254::G1AffinePoint>(max_point_num);
   std::vector<bn254::Fr> scalars = base::CreateVectorParallel(
